feat: first step of MVVM
The idea is to convert the application to use the MVVM pattern. For this reason, two new classes have been added (ViewModel, Repository). This is the initial commit, more work coming.

This commit add two files, MovieViewModel and MovieRepository. It also
modifies the current files to use the MovieViewModel file.

Relates: #1